+++ src/BatteryMonitor.h
@@ -1,20 +1,15 @@
 #ifndef __BATTERY_CONTROLLER_H__
 #define __BATTERY_CONTROLLER_H__
 
+#include <Logger.h>
 #include "config.h"
 #include "Buzzer.h"
 #ifdef CANBUS_ENABLED
  #include "CanBus.h"
 #endif
-
-#ifdef BATTERY_BAR
- #include <Adafruit_NeoPixel.h>
- #ifndef BATTERY_BAR_PIN
-  #define BATTERY_BAR_PIN 5 // default PIN
- #endif //BATTERY_BAR_PIN
- extern Adafruit_NeoPixel batPixels;
-#endif //BATTERY_BAR
-
+#ifdef LIGHT_BAR_ENABLED
+ #include "LightBarController.h"
+#endif
 
 class BatteryMonitor {
     public:
@@ -30,8 +25,6 @@ class BatteryMonitor {
 #ifdef CANBUS_ENABLED
         CanBus::VescData *vescData;
 #endif
-        void updateBatteryBar(float voltage);
-        int calcVal(int value);
         int smoothAnalogReading();
         void updateCurrentArray(double value);
         double getAverageCurrent();