+++ src/OTA.cpp
@@ -7,45 +7,118 @@ bool updateFlag = false;
 bool readyFlag = false;
 int bytesReceived = 0;
 int timesWritten = 0;
+uint32_t frameNumber = 0;
 
 OTAUpdater::OTAUpdater() {
 }
 
 void OTAUpdater::setup() {
   begin("rESCue OTA Updates");
+  Buzzer::getInstance()->playSound(RTTTL_MELODIES::SIMPLE_BEEP_SCALE_UP);
+  while (Buzzer::getInstance()->isPlayingSound()) {
+    ;
+  }
+
+	// Get Partitionsizes
+	size_t ul;
+	esp_partition_iterator_t _mypartiterator;
+	const esp_partition_t *_mypart;
+	ul = spi_flash_get_chip_size();
+	Serial.print("\nFlash chip size: ");
+	Serial.println(ul);
+	Serial.println("Partition table:");
+	_mypartiterator = esp_partition_find(ESP_PARTITION_TYPE_APP, ESP_PARTITION_SUBTYPE_ANY, NULL);
+
+  while (_mypartiterator != NULL) {
+		_mypart = esp_partition_get(_mypartiterator);
+		printf("Type: %02x SubType %02x Address 0x%06X Size 0x%06X Encryption %i Label %s\r\n", _mypart->type, _mypart->subtype, _mypart->address, _mypart->size, _mypart->encrypted, _mypart->label);
+		_mypartiterator = esp_partition_next(_mypartiterator);
+	}
+
+  _mypart = esp_ota_get_boot_partition();
+  printf("Current active partition is %s\r\n", _mypart->label);
+
+}
+
+NimBLEUUID OTAUpdater::getConfCharacteristicsUuid() {
+  return pESPOTAConfCharacteristic->getUUID();
 }
 
 void OTACallback::onWrite(BLECharacteristic *pCharacteristic) {
   std::string rxData = pCharacteristic->getValue();
-  if (!updateFlag) { //If it's the first packet of OTA since bootup, begin OTA
-    AppConfiguration::getInstance()->config.otaUpdateActive = 0;
-    AppConfiguration::getInstance()->savePreferences();
 
-    Serial.println("BeginOTA");
-    esp_ota_begin(esp_ota_get_next_update_partition(NULL), OTA_SIZE_UNKNOWN, &otaHandler);
-    updateFlag = true;
-  }
-  if (_p_ble != NULL) {
-    if (rxData.length() > 0) {
-      esp_ota_write(otaHandler, rxData.c_str(), rxData.length());
-      if (rxData.length() != FULL_PACKET) {
-        esp_ota_end(otaHandler);
-        Serial.println("EndOTA");
-        if (ESP_OK == esp_ota_set_boot_partition(esp_ota_get_next_update_partition(NULL))) {
-          delay(2000);
-          esp_restart();
-        }
-        else {
-          Serial.println("Upload Error");
+  if(pCharacteristic->getUUID().equals(_p_ble->getConfCharacteristicsUuid())) {
+      //Serial.println("Write config");
+      //Serial.println("String " + String(rxData.c_str()));
+      //Serial.println("Length " + String(rxData.length()));
+      
+      std::string str(rxData.c_str());
+      std::string::size_type middle = str.find('='); // Find position of ','
+      std::string key = "";
+      std::string value = "";      
+      if(middle != std::string::npos) {
+        key = str.substr(0, middle);
+        value = str.substr(middle + 1, str.size() - (middle + 1));
+      }
+
+      Serial.println(String(key.c_str()) + String("=") + String(value.c_str()));
+
+      if(key == "numberPixelLight") {
+        AppConfiguration::getInstance()->config.numberPixelLight = atoi(value.c_str());
+      }
+      if(key == "numberPixelBatMon") {
+        AppConfiguration::getInstance()->config.numberPixelBatMon = atoi(value.c_str());
+      }
+      if(key == "vescId") {
+        AppConfiguration::getInstance()->config.vescId = atoi(value.c_str());
+      } 
+      if(key == "authToken") {
+        AppConfiguration::getInstance()->config.authToken = value.c_str();
+      }
+      if(key == "save") {
+        AppConfiguration::getInstance()->config.otaUpdateActive = false;
+        AppConfiguration::getInstance()->savePreferences();
+        delay(100);
+    }
+  } else {
+    if (!updateFlag) { //If it's the first packet of OTA since bootup, begin OTA
+      Serial.println("\nBeginOTA");
+      const esp_partition_t *partition = esp_ota_get_next_update_partition(NULL);
+      Serial.println("Selected OTA partiton:");
+      Serial.println("partition label:" + String(partition->label));
+      Serial.println("partition size:" + String(partition->size));
+      esp_ota_begin(partition, OTA_SIZE_UNKNOWN, &otaHandler);
+      updateFlag = true;
+    }
+    if (_p_ble != NULL) {
+      if (rxData.length() > 0) {
+        Serial.print("Got frame " + String(frameNumber));
+        esp_ota_write(otaHandler, rxData.c_str(), rxData.length());
+        if (rxData.length() != FULL_PACKET) {
+          esp_ota_end(otaHandler);
+          Serial.println("\nEndOTA");
+          const esp_partition_t *partition = esp_ota_get_next_update_partition(NULL);
+          if (ESP_OK == esp_ota_set_boot_partition(partition)) {
+            Serial.println("Activate partiton:");
+            Serial.println("partition label:" + String(partition->label));
+            Serial.println("partition size:" + String(partition->size));
+            AppConfiguration::getInstance()->config.otaUpdateActive = 0;
+            AppConfiguration::getInstance()->savePreferences();
+            delay(2000);
+            esp_restart();
+          } else {
+            Serial.println("Upload Error");
+          }
         }
       }
     }
-  }
 
-  uint8_t txData[5] = {1, 2, 3, 4, 5};
-  //delay(1000);
-  pCharacteristic->setValue((uint8_t*)txData, 5);
-  pCharacteristic->notify();
+    delay(5); // needed to give BLE stack some time
+    uint8_t txdata[4] = { (uint8_t)(frameNumber >> 24), (uint8_t)(frameNumber >> 16), (uint8_t)(frameNumber >> 8), (uint8_t)frameNumber};
+    pCharacteristic->setValue((uint8_t *) txdata, 4);
+    pCharacteristic->notify();
+    Serial.println(", Ack. frame no. " + String(frameNumber++));
+  }
 }
 
 bool OTAUpdater::begin(const char* localName = "UART Service") {
@@ -66,6 +139,10 @@ bool OTAUpdater::begin(const char* localName = "UART Service") {
                                        NIMBLE_PROPERTY::READ
                                      );
 
+  pESPOTAConfCharacteristic = pESPOTAService->createCharacteristic(
+                         CHARACTERISTIC_UUID_CONF,
+                         NIMBLE_PROPERTY::NOTIFY | NIMBLE_PROPERTY::WRITE
+                       );
   pVersionCharacteristic = pService->createCharacteristic(
                              CHARACTERISTIC_UUID_HW_VERSION,
                             NIMBLE_PROPERTY::READ
@@ -76,6 +153,7 @@ bool OTAUpdater::begin(const char* localName = "UART Service") {
                          NIMBLE_PROPERTY::NOTIFY | NIMBLE_PROPERTY::WRITE
                        );
 
+  pESPOTAConfCharacteristic->setCallbacks(new OTACallback(this));
   pOtaCharacteristic->setCallbacks(new OTACallback(this));
 
   // Start the service(s)